[Refactor] #4 Badge파일로 분리
